# srv-espaco/index.js
const express = require('express');
const axios = require('axios');
const fs = require('fs');
const swaggerUi = require('swagger-ui-express');
const yaml = require('js-yaml');
const path = require('path');
const crypto = require('crypto');

const app = express();
const PORT = 3002;

// Middleware para JSON
app.use(express.json());

// --- Configuração do Swagger ---
try {
  const fileContents = fs.readFileSync(path.join(__dirname, 'openapi.yaml'), 'utf8');
  const swaggerDocument = yaml.load(fileContents);
  app.use('/docs', swaggerUi.serve, swaggerUi.setup(swaggerDocument));
  console.log(`Swagger disponível em http://localhost:${PORT}/docs`);
} catch (e) {
  console.error("Erro ao carregar openapi.yaml:", e);
}

// --- Configuração do "Banco de Dados" (JSON) ---
const DB_DIR = path.join(__dirname, 'database');
const DB_FILE = path.join(DB_DIR, 'dados-espaco.json');

// Garante que a pasta e o arquivo existam ao iniciar
if (!fs.existsSync(DB_DIR)) {
  fs.mkdirSync(DB_DIR, { recursive: true });
}
if (!fs.existsSync(DB_FILE)) {
  fs.writeFileSync(DB_FILE, '[]', 'utf8'); // Inicia com array vazio
}

// Funções auxiliares de persistência
function lerBanco() {
  try {
    const data = fs.readFileSync(DB_FILE, 'utf8');
    return JSON.parse(data);
  } catch (error) {
    console.error("Erro ao ler banco:", error);
    return [];
  }
}

function salvarBanco(dados) {
  try {
    fs.writeFileSync(DB_FILE, JSON.stringify(dados, null, 2), 'utf8');
  } catch (error) {
    console.error("Erro ao salvar banco:", error);
  }
}

// --- Rotas da API ---

// 1. GET /espacos
app.get('/espacos', (req, res) => {
  const espacos = lerBanco();
  res.status(200).json(espacos);
});

// 2. GET /espacos/:id
app.get('/espacos/:id', (req, res) => {
  const espacos = lerBanco();
  const espaco = espacos.find(e => e.id === req.params.id);
  
  if (!espaco) {
    return res.status(404).json({ error: 'Espaço não encontrado.' });
  }
  
  res.status(200).json(espaco);
});

// 3. POST /espacos
app.post('/espacos', (req, res) => {
  const { nome, capacidade, preco, fotos } = req.body;

  if (!nome || !capacidade || !preco) {
    return res.status(400).json({ error: 'Campos obrigatórios: nome, capacidade, preco.' });
  }

  const espacos = lerBanco();
  
  const novoEspaco = {
    id: crypto.randomUUID(),
    nome,
    capacidade,
    preco,
    fotos: fotos || []
  };

  espacos.push(novoEspaco);
  salvarBanco(espacos); // Persiste no arquivo

  res.status(201).json(novoEspaco);
});

// 4. PUT /espacos/:id
app.put('/espacos/:id', (req, res) => {
  const espacos = lerBanco();
  const index = espacos.findIndex(e => e.id === req.params.id);
  
  if (index === -1) {
    return res.status(404).json({ error: 'Espaço não encontrado.' });
  }

  const { nome, capacidade, preco, fotos } = req.body;

  // Substitui o objeto mantendo o ID original
  espacos[index] = {
    id: req.params.id,
    nome,
    capacidade,
    preco,
    fotos: fotos || []
  };

  salvarBanco(espacos);
  res.status(200).json(espacos[index]);
});

// 5. PATCH /espacos/:id
app.patch('/espacos/:id', (req, res) => {
  const espacos = lerBanco();
  const index = espacos.findIndex(e => e.id === req.params.id);
  
  if (index === -1) {
    return res.status(404).json({ error: 'Espaço não encontrado.' });
  }

  // Atualiza apenas os campos enviados
  const espacoAtualizado = { ...espacos[index], ...req.body };
  // Garante que o ID não foi alterado
  espacoAtualizado.id = req.params.id;
  
  espacos[index] = espacoAtualizado;
  
  salvarBanco(espacos);
  res.status(200).json(espacoAtualizado);
});

// 6. DELETE /espacos/:id
app.delete('/espacos/:id', (req, res) => {
  let espacos = lerBanco();
  const index = espacos.findIndex(e => e.id === req.params.id);
  
  if (index === -1) {
    return res.status(404).json({ error: 'Espaço não encontrado.' });
  }

  espacos.splice(index, 1);
  salvarBanco(espacos);
  
  res.status(204).send();
});

app.listen(PORT, () => {
  console.log(`Serviço de Espaços rodando na porta ${PORT}`);
  console.log(`Banco de dados: ${DB_FILE}`);
});